services/x_alpha/requirements.txt
# X-Alpha 智能舆情服务依赖

# Web 框架
fastapi>=0.104.0
uvicorn>=0.24.0

# HTTP 客户端
httpx>=0.25.0

# AI API 客户端
openai>=1.3.0

# HTML 解析
beautifulsoup4>=4.12.0

# 数据验证
pydantic>=2.5.0

# 环境变量
python-dotenv>=1.0.0

# CLI
click>=8.1.0
